requirements.txt
requests>=2.31.0
urllib3>=1.26.0
GitPython>=3.1.40
click>=8.1.7
pyyaml>=6.0.1
jinja2>=3.1.2

# Development and testing dependencies
pytest>=7.4.0
pytest-mock>=3.11.1
responses>=0.23.1
pytest-cov>=4.1.0